+++ src/aat/java/uk/gov/hmcts/cmc/claimstore/tests/BaseTest.java
@@ -6,7 +6,6 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.context.junit4.SpringRunner;
-import uk.gov.hmcts.cmc.claimstore.tests.Bootstrap;
 import uk.gov.hmcts.cmc.claimstore.processors.JsonMapper;
 
 @RunWith(SpringRunner.class)
